+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v20260523-0
 
 * [#80](https://github.com/RobertWesner/YouTube-Play-All/issues/80) Use new `--yt-sys` CSS variables with fallback to old `--yt-spec`.
+* [#81](https://github.com/RobertWesner/YouTube-Play-All/issues/81) Improved detection of currently selected button, working around YouTube constantly unselecting it and erasing modifications.
 
 
 ## v20260508-0